Be event

See merge request s05-webmobile1-sub3/S05P13C103!59

+++ backend/src/main/java/com/ssafy/api/response/dto/User/LivewithUser.java
@@ -8,10 +8,10 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class LivewithUser {
-    Long livepk;
-    Long productpk;
+    Integer livepk;
+    Integer productpk;
     String livetitle;
-    Long liveviewercount;
+    Integer liveviewercount;
     String userid;
     String usernickname;
     LocalDateTime usercreatedat;
